IPCam_Controller/requirements.txt
loguru==0.7.0
mysql-connector-python==8.0.33
numpy==1.24.3
opencv-python-headless==4.7.0.72
Pillow==9.5.0
protobuf==3.20.3
PyYAML==6.0
